+++ test/audits/forms.test.js
@@ -24,12 +24,13 @@ import {
 import { MockContextBuilder } from '../shared.js';
 import formVitalsData from '../fixtures/formvitalsdata.json' with { type: 'json' };
 import testData from '../fixtures/high-form-views-low-conversions.js';
-import convertToOpportunity from '../../src/forms-opportunities/opportunityHandler.js';
+import createLowConversionOpportunities from '../../src/forms-opportunities/oppty-handlers/low-conversion-handler.js';
 import { shouldExcludeForm } from '../../src/forms-opportunities/utils.js';
 import expectedFormVitalsData from '../fixtures/expectedformvitalsdata.json' with { type: 'json' };
 import expectedFormSendToScraperData from '../fixtures/expectedformsendtoscraperdata.json' with { type: 'json' };
 import formScrapeData from '../fixtures/formscrapedata.js';
-import highPageViewsLowFormNavOpportunity from '../../src/forms-opportunities/highPageViewsLowFormNavOpportunity.js';
+import createLowNavigationOpportunities from '../../src/forms-opportunities/oppty-handlers/low-navigation-handler.js';
+import { FORM_OPPORTUNITY_TYPES } from '../../src/forms-opportunities/constants.js';
 
 use(sinonChai);
 
@@ -90,7 +91,7 @@ describe('opportunities handler method', () => {
       getId: () => 'opportunity-id',
       setAuditId: sinon.stub(),
       save: sinon.stub(),
-      getType: () => 'high-form-views-low-conversions',
+      getType: () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION,
       setData: sinon.stub(),
       setGuidance: sinon.stub(),
       getData: sinon.stub().returns({
@@ -132,9 +133,9 @@ describe('opportunities handler method', () => {
   });
 
   it('should create new forms opportunity', async () => {
-    formsOppty.getType = () => 'high-form-views-low-conversions';
+    formsOppty.getType = () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION;
     dataAccessStub.Opportunity.create = sinon.stub().returns(formsOppty);
-    await convertToOpportunity(auditUrl, auditData, undefined, context);
+    await createLowConversionOpportunities(auditUrl, auditData, undefined, context);
     expect(dataAccessStub.Opportunity.create).to.be.callCount(5);
     expect(dataAccessStub.Opportunity.create).to.be.calledWith(testData.opportunityData);
     // with empty guidance due to no scraping
@@ -142,9 +143,10 @@ describe('opportunities handler method', () => {
   });
 
   it('should create new forms opportunity with scraped data available', async () => {
-    formsOppty.getType = () => 'high-form-views-low-conversions';
+    formsOppty.getType = () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION;
     dataAccessStub.Opportunity.create = sinon.stub().returns(formsOppty);
-    await convertToOpportunity(auditUrl, auditData, formScrapeData.scrapeData1, context);
+    const { scrapeData1: scrapeData } = formScrapeData;
+    await createLowConversionOpportunities(auditUrl, auditData, scrapeData, context);
     // with BTF guidance
     // expect(dataAccessStub.Opportunity.create).to.not.have.been.called;
     expect(logStub.info).to.be.calledWith('Successfully synced Opportunity for site: site-id and high-form-views-low-conversions audit type.');
@@ -152,18 +154,26 @@ describe('opportunities handler method', () => {
 
   it('should create new forms opportunity with scraped data available not matched', async () => {
     const { auditData2 } = testData;
-    formsOppty.getType = () => 'high-form-views-low-conversions';
+    formsOppty.getType = () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION;
     dataAccessStub.Opportunity.create = sinon.stub().returns(formsOppty);
-    await convertToOpportunity(auditUrl, auditData2, formScrapeData.scrapeData2, context);
+    const { scrapeData2: scrapeData } = formScrapeData;
+    await createLowConversionOpportunities(auditUrl, auditData2, scrapeData, context);
     expect(dataAccessStub.Opportunity.create).to.be.calledWith(testData.opportunityData2);
     // with empty guidance due to scrapedStatus = false
     expect(logStub.info).to.be.calledWith('Successfully synced Opportunity for site: site-id and high-form-views-low-conversions audit type.');
   });
 
   it('should create new forms opportunity with scraped data available with all field labels containing search', async () => {
-    formsOppty.getType = () => 'high-form-views-low-conversions';
+    formsOppty.getType = () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION;
     dataAccessStub.Opportunity.create = sinon.stub().returns(formsOppty);
-    await convertToOpportunity(auditUrl, auditData, formScrapeData.scrapeData3, context);
+    const { scrapeData3: scrapeData } = formScrapeData;
+    const { auditDataOpportunitiesWithSearchFields } = testData;
+    await createLowConversionOpportunities(
+      auditUrl,
+      auditDataOpportunitiesWithSearchFields,
+      scrapeData,
+      context,
+    );
     const expectedOpportunityData = { ...testData.opportunityData3 };
     // with large form guidance
     expectedOpportunityData.data.scrapedStatus = true;
@@ -173,17 +183,25 @@ describe('opportunities handler method', () => {
 
   it('should create new forms opportunity with scraped data available and matched with Generic guidance', async () => {
     const { auditData2 } = testData;
-    formsOppty.getType = () => 'high-form-views-low-conversions';
+    formsOppty.getType = () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION;
     dataAccessStub.Opportunity.create = sinon.stub().returns(formsOppty);
-    await convertToOpportunity(auditUrl, auditData2, formScrapeData.scrapeData4, context);
+    const { scrapeData4: scrapeData } = formScrapeData;
+    await createLowConversionOpportunities(auditUrl, auditData2, scrapeData, context);
     expect(dataAccessStub.Opportunity.create).to.be.calledWith(testData.opportunityData4);
+
     // with Generic guidance
     expect(logStub.info).to.be.calledWith('Successfully synced Opportunity for site: site-id and high-form-views-low-conversions audit type.');
   });
 
   it('should use existing opportunity', async () => {
     dataAccessStub.Opportunity.allBySiteIdAndStatus.resolves([formsOppty]);
-    await convertToOpportunity(auditUrl, auditData, undefined, context);
+    const { auditDataWithExistingOppty } = testData;
+    await createLowConversionOpportunities(
+      auditUrl,
+      auditDataWithExistingOppty,
+      undefined,
+      context,
+    );
     expect(formsOppty.save).to.be.callCount(1);
     expect(logStub.info).to.be.calledWith('Successfully synced Opportunity for site: site-id and high-form-views-low-conversions audit type.');
   });
@@ -191,7 +209,7 @@ describe('opportunities handler method', () => {
   it('should throw error if fetching opportunity fails', async () => {
     dataAccessStub.Opportunity.allBySiteIdAndStatus.rejects(new Error('some-error'));
     try {
-      await convertToOpportunity(auditUrl, auditData, undefined, context);
+      await createLowConversionOpportunities(auditUrl, auditData, undefined, context);
     } catch (err) {
       expect(err.message).to.equal('Failed to fetch opportunities for siteId site-id: some-error');
     }
@@ -202,11 +220,26 @@ describe('opportunities handler method', () => {
     dataAccessStub.Opportunity.allBySiteIdAndStatus.returns([]);
     dataAccessStub.Opportunity.create = sinon.stub().rejects(new Error('some-error'));
     try {
-      await convertToOpportunity(auditUrl, auditData, undefined, context);
+      await createLowConversionOpportunities(auditUrl, auditData, undefined, context);
     } catch (err) {
       expect(err.message).to.equal('Failed to create Forms opportunity for siteId site-id: some-error');
     }
     expect(logStub.error).to.be.calledWith('Creating Forms opportunity for siteId site-id failed with error: some-error');
+  });
+
+  it('should create new forms opportunity with device wise metrics and traffic ', async () => {
+    formsOppty.getType = () => 'high-form-views-low-conversions';
+    dataAccessStub.Opportunity.create = sinon.stub().returns(formsOppty);
+    const { auditDataWithTrafficMetrics } = testData;
+    await createLowConversionOpportunities(
+      auditUrl,
+      auditDataWithTrafficMetrics,
+      undefined,
+      context,
+    );
+    expect(dataAccessStub.Opportunity.create).to.be.callCount(1);
+    expect(dataAccessStub.Opportunity.create).to.be.calledWith(testData.opportunityData5);
+    expect(logStub.info).to.be.calledWith('Successfully synced Opportunity for site: site-id and high-form-views-low-conversions audit type.');
   });
 });
 
@@ -237,7 +270,7 @@ describe('audit and send scraping step', () => {
         getId: () => 'opportunity-id',
         setAuditId: sinon.stub(),
         save: sinon.stub(),
-        getType: () => 'high-form-views-low-conversions',
+        getType: () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION,
       },
       finalUrl: 'www.example.com',
     })
@@ -271,7 +304,7 @@ describe('process opportunity step', () => {
     getId: () => 'opportunity-id',
     setAuditId: sinon.stub(),
     save: sinon.stub(),
-    getType: () => 'high-form-views-low-conversions',
+    getType: () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION,
   };
 
   let dataAccessStub = {
@@ -286,7 +319,7 @@ describe('process opportunity step', () => {
       getId: () => 'opportunity-id',
       setAuditId: sinon.stub(),
       save: sinon.stub(),
-      getType: () => 'high-form-views-low-conversions',
+      getType: () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION,
     };
 
     dataAccessStub = {
@@ -330,7 +363,7 @@ describe('process opportunity step', () => {
           getId: 'opportunity-id',
           setAuditId: sinon.stub(),
           save: sinon.stub(),
-          getType: 'high-form-views-low-conversions',
+          getType: FORM_OPPORTUNITY_TYPES.LOW_CONVERSION,
         },
       },
       env: {
@@ -384,7 +417,7 @@ describe('process opportunity step', () => {
     };
 
     context.s3Client.send.resolves(mockFileResponse);
-    formsOppty.getType = () => 'high-form-views-low-conversions';
+    formsOppty.getType = () => FORM_OPPORTUNITY_TYPES.LOW_CONVERSION;
     dataAccessStub.Opportunity.create = sinon.stub().returns(formsOppty);
 
     const result = await processOpportunityStep(context);
@@ -394,7 +427,7 @@ describe('process opportunity step', () => {
   });
 });
 
-describe('highPageViewsLowFormNavOpportunity handler method', () => {
+describe('createLowNavigationOpportunities handler method', () => {
   let logStub;
   let dataAccessStub;
   let auditData;
@@ -409,7 +442,7 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
       getId: () => 'opportunity-id',
       setAuditId: sinon.stub(),
       save: sinon.stub(),
-      getType: () => 'high-page-views-low-form-nav',
+      getType: () => FORM_OPPORTUNITY_TYPES.LOW_NAVIGATION,
       setData: sinon.stub(),
       setGuidance: sinon.stub(),
       getData: sinon.stub().returns({
@@ -451,7 +484,7 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
       siteId: 'site-id',
       auditId: 'audit-id',
       runbook: 'https://adobe.sharepoint.com/:w:/s/AEM_Forms/ETCwSsZJzRJIuPqnC_jZFhgBsW29GijIgk9C6-GpkQ16xg?e=dNYZhD',
-      type: 'high-page-views-low-form-nav',
+      type: FORM_OPPORTUNITY_TYPES.LOW_NAVIGATION,
       origin: 'AUTOMATION',
       title: 'Form has low views',
       description: 'The form has low views due to low navigations in the page containing its CTA',
@@ -461,18 +494,42 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
       data: {
         form: 'https://www.surest.com/newsletter',
         screenshot: '',
-        trackedFormKPIName: 'Conversion Rate',
-        trackedFormKPIValue: null,
+        trackedFormKPIName: 'Form Views',
+        trackedFormKPIValue: 300,
         formViews: 300,
         pageViews: 8670,
         samples: 8670,
         scrapedStatus: false,
         metrics: [
           {
-            type: 'conversionRate',
+            type: 'formViews',
             device: '*',
             value: {
-              page: null,
+              page: 300,
+            },
+          },
+          {
+            type: 'formViews',
+            device: 'mobile',
+            value: {
+              page: 300,
+            },
+          },
+          {
+            type: 'formViews',
+            device: 'desktop',
+            value: {
+              page: 0,
+            },
+          },
+          {
+            type: 'traffic',
+            device: '*',
+            value: {
+              paid: 4670,
+              total: 8670,
+              earned: 2000,
+              owned: 2000,
             },
           },
         ],
@@ -493,10 +550,10 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
       },
     };
 
-    formsCTAOppty.getType = () => 'high-page-views-low-form-nav';
+    formsCTAOppty.getType = () => FORM_OPPORTUNITY_TYPES.LOW_NAVIGATION;
     dataAccessStub.Opportunity.create = sinon.stub().returns(formsCTAOppty);
 
-    await highPageViewsLowFormNavOpportunity(auditUrl, auditData, undefined, context);
+    await createLowNavigationOpportunities(auditUrl, auditData, undefined, context);
 
     const actualCall = dataAccessStub.Opportunity.create.getCall(0).args[0];
     expect(actualCall).to.deep.equal(expectedOpportunityData);
@@ -506,7 +563,7 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
   it('should use existing high page views low form navigation opportunity', async () => {
     dataAccessStub.Opportunity.allBySiteIdAndStatus.resolves([formsCTAOppty]);
 
-    await highPageViewsLowFormNavOpportunity(auditUrl, auditData, undefined, context);
+    await createLowNavigationOpportunities(auditUrl, auditData, undefined, context);
 
     expect(formsCTAOppty.save).to.be.calledOnce;
     expect(logStub.info).to.be.calledWith('Successfully synced Opportunity for site: site-id and high page views low form nav audit type.');
@@ -516,7 +573,7 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
     dataAccessStub.Opportunity.allBySiteIdAndStatus.rejects(new Error('some-error'));
 
     try {
-      await highPageViewsLowFormNavOpportunity(auditUrl, auditData, undefined, context);
+      await createLowNavigationOpportunities(auditUrl, auditData, undefined, context);
     } catch (err) {
       expect(err.message).to.equal('Failed to fetch opportunities for siteId site-id: some-error');
     }
@@ -529,7 +586,7 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
     dataAccessStub.Opportunity.create = sinon.stub().rejects(new Error('some-error'));
 
     try {
-      await highPageViewsLowFormNavOpportunity(auditUrl, auditData, undefined, context);
+      await createLowNavigationOpportunities(auditUrl, auditData, undefined, context);
     } catch (err) {
       expect(err.message).to.equal('Failed to create Forms opportunity for high page views low form nav for siteId site-id: some-error');
     }
@@ -540,7 +597,7 @@ describe('highPageViewsLowFormNavOpportunity handler method', () => {
   it('should handle empty form vitals data', async () => {
     auditData.auditResult.formVitals = [];
 
-    await highPageViewsLowFormNavOpportunity(auditUrl, auditData, undefined, context);
+    await createLowNavigationOpportunities(auditUrl, auditData, undefined, context);
 
     expect(dataAccessStub.Opportunity.create).to.not.be.called;
     expect(logStub.info).to.be.calledWith('Successfully synced Opportunity for site: site-id and high page views low form nav audit type.');
@@ -560,16 +617,6 @@ describe('isSearchForm', () => {
 
   it('should return true for form with unsubscribe class', () => {
     const scrapedFormData = { classList: ['unsubscribe'] };
-    expect(shouldExcludeForm(scrapedFormData)).to.be.true;
-  });
-
-  it('should return true for form with action ending in search.html', () => {
-    const scrapedFormData = { action: 'https://example.com/search.html' };
-    expect(shouldExcludeForm(scrapedFormData)).to.be.true;
-  });
-
-  it('should return true for form with all field labels containing search', () => {
-    const scrapedFormData = { fieldsLabels: ['Search', 'Advanced Search'] };
     expect(shouldExcludeForm(scrapedFormData)).to.be.true;
   });